+++ src/bin/cli.rs
@@ -1,19 +1,8 @@
 extern crate clap;
 
-use std::cmp::min;
-use std::collections::HashMap;
-use std::fs::File;
-use std::io::{self, Read, BufRead};
-use std::process;
-
 use clap::{AppSettings, Arg, ArgMatches};
-use toml;
 
-use cargo::{self, Config, CargoError, CliResult, CliError};
-use cargo::core::{Source, SourceId, GitReference, Package, Verbosity};
-use cargo::ops::{self, CompileMode, OutputMetadataOptions};
-use cargo::sources::{GitSource, RegistrySource};
-use cargo::util::{ToUrl, CargoResultExt};
+use cargo::{self, Config, CliResult};
 
 use super::list_commands;
 use super::commands;
@@ -63,8 +52,6 @@ pub fn main(config: &mut Config) -> CliResult {
     }
 
     if args.subcommand_name().is_none() {
-        cli().print_help()?;
-        return Ok(());
     }
 
     execute_subcommand(config, args)
@@ -80,481 +67,28 @@ fn execute_subcommand(config: &mut Config, args: ArgMatches) -> CliResult {
         &args.values_of_lossy("unstable-features").unwrap_or_default(),
     )?;
 
-    match args.subcommand() {
-        ("bench", Some(args)) => {
-            let ws = args.workspace(config)?;
-            let mut compile_opts = args.compile_options(config, CompileMode::Bench)?;
-            compile_opts.release = true;
-
-            let ops = ops::TestOptions {
-                no_run: args.is_present("no-run"),
-                no_fail_fast: args.is_present("no-fail-fast"),
-                only_doc: false,
-                compile_opts,
-            };
-
-            let mut bench_args = vec![];
-            bench_args.extend(args.value_of("BENCHNAME").into_iter().map(|s| s.to_string()));
-            bench_args.extend(args.values_of("args").unwrap_or_default().map(|s| s.to_string()));
-
-            let err = ops::run_benches(&ws, &ops, &bench_args)?;
-            match err {
-                None => Ok(()),
-                Some(err) => {
-                    Err(match err.exit.as_ref().and_then(|e| e.code()) {
-                        Some(i) => CliError::new(format_err!("bench failed"), i),
-                        None => CliError::new(err.into(), 101)
-                    })
-                }
-            }
+    let (cmd, args) = match args.subcommand() {
+        (cmd, Some(args)) => (cmd, args),
+        _ => {
+            cli().print_help()?;
+            return Ok(());
         }
-        ("build", Some(args)) => {
-            let mut ws = args.workspace(config)?;
-            let compile_opts = args.compile_options(config, CompileMode::Build)?;
-            ops::compile(&ws, &compile_opts)?;
-            Ok(())
-        }
-        ("check", Some(args)) => {
-            let ws = args.workspace(config)?;
-            let test = match args.value_of("profile") {
-                Some("test") => true,
-                None => false,
-                Some(profile) => {
-                    let err = format_err!("unknown profile: `{}`, only `test` is \
-                                       currently supported", profile);
-                    return Err(CliError::new(err, 101));
-                }
-            };
-            let mode = CompileMode::Check { test };
-            let compile_opts = args.compile_options(config, mode)?;
-            ops::compile(&ws, &compile_opts)?;
-            Ok(())
-        }
-        ("clean", Some(args)) => {
-            let ws = args.workspace(config)?;
-            let opts = ops::CleanOptions {
-                config,
-                spec: values(args, "package"),
-                target: args.target(),
-                release: args.is_present("release"),
-            };
-            ops::clean(&ws, &opts)?;
-            Ok(())
-        }
-        ("doc", Some(args)) => {
-            let ws = args.workspace(config)?;
-            let mode = ops::CompileMode::Doc { deps: !args.is_present("no-deps") };
-            let compile_opts = args.compile_options(config, mode)?;
-            let doc_opts = ops::DocOptions {
-                open_result: args.is_present("open"),
-                compile_opts,
-            };
-            ops::doc(&ws, &doc_opts)?;
-            Ok(())
-        }
-        ("fetch", Some(args)) => {
-            let ws = args.workspace(config)?;
-            ops::fetch(&ws)?;
-            Ok(())
-        }
-        ("generate-lockfile", Some(args)) => {
-            let ws = args.workspace(config)?;
-            ops::generate_lockfile(&ws)?;
-            Ok(())
-        }
-        ("git-checkout", Some(args)) => {
-            let url = args.value_of("url").unwrap().to_url()?;
-            let reference = args.value_of("reference").unwrap();
+    };
 
-            let reference = GitReference::Branch(reference.to_string());
-            let source_id = SourceId::for_git(&url, reference)?;
-
-            let mut source = GitSource::new(&source_id, config)?;
-
-            source.update()?;
-
-            Ok(())
-        }
-        ("init", Some(args)) => {
-            let opts = args.new_options()?;
-            ops::init(&opts, config)?;
-            config.shell().status("Created", format!("{} project", opts.kind))?;
-            Ok(())
-        }
-        ("install", Some(args)) => {
-            let mut compile_opts = args.compile_options(config, CompileMode::Build)?;
-            compile_opts.release = !args.is_present("debug");
-
-            let krates = args.values_of("crate").unwrap_or_default().collect::<Vec<_>>();
-
-            let source = if let Some(url) = args.value_of("git") {
-                let url = url.to_url()?;
-                let gitref = if let Some(branch) = args.value_of("branch") {
-                    GitReference::Branch(branch.to_string())
-                } else if let Some(tag) = args.value_of("tag") {
-                    GitReference::Tag(tag.to_string())
-                } else if let Some(rev) = args.value_of("rev") {
-                    GitReference::Rev(rev.to_string())
-                } else {
-                    GitReference::Branch("master".to_string())
-                };
-                SourceId::for_git(&url, gitref)?
-            } else if let Some(path) = args.value_of("path") {
-                SourceId::for_path(&config.cwd().join(path))?
-            } else if krates.is_empty() {
-                SourceId::for_path(config.cwd())?
-            } else {
-                SourceId::crates_io(config)?
-            };
-
-            let version = args.value_of("version");
-            let root = args.value_of("root");
-
-            if args.is_present("list") {
-                ops::install_list(root, config)?;
-            } else {
-                ops::install(root, krates, &source, version, &compile_opts, args.is_present("force"))?;
-            }
-            Ok(())
-        }
-        ("locate-project", Some(args)) => {
-            let root = args.root_manifest(config)?;
-
-            let root = root.to_str()
-                .ok_or_else(|| format_err!("your project path contains characters \
-                                            not representable in Unicode"))
-                .map_err(|e| CliError::new(e, 1))?
-                .to_string();
-
-            #[derive(Serialize)]
-            pub struct ProjectLocation {
-                root: String
-            }
-
-            let location = ProjectLocation { root };
-
-            cargo::print_json(&location);
-            Ok(())
-        }
-        ("login", Some(args)) => {
-            let registry = args.registry(config)?;
-
-            let token = match args.value_of("token") {
-                Some(token) => token.to_string(),
-                None => {
-                    let host = match registry {
-                        Some(ref _registry) => {
-                            return Err(format_err!("token must be provided when \
-                                            --registry is provided.").into());
-                        }
-                        None => {
-                            let src = SourceId::crates_io(config)?;
-                            let mut src = RegistrySource::remote(&src, config);
-                            src.update()?;
-                            let config = src.config()?.unwrap();
-                            args.value_of("host").map(|s| s.to_string())
-                                .unwrap_or(config.api.unwrap())
-                        }
-                    };
-                    println!("please visit {}me and paste the API Token below", host);
-                    let mut line = String::new();
-                    let input = io::stdin();
-                    input.lock().read_line(&mut line).chain_err(|| {
-                        "failed to read stdin"
-                    }).map_err(CargoError::from)?;
-                    line.trim().to_string()
-                }
-            };
-
-            ops::registry_login(config, token, registry)?;
-            Ok(())
-        }
-        ("metadata", Some(args)) => {
-            let ws = args.workspace(config)?;
-
-            let version = match args.value_of("format-version") {
-                None => {
-                    config.shell().warn("\
-                        please specify `--format-version` flag explicitly \
-                        to avoid compatibility problems"
-                    )?;
-                    1
-                }
-                Some(version) => version.parse().unwrap(),
-            };
-
-            let options = OutputMetadataOptions {
-                features: values(args, "features"),
-                all_features: args.is_present("all-features"),
-                no_default_features: args.is_present("no-default-features"),
-                no_deps: args.is_present("no-deps"),
-                version,
-            };
-
-            let result = ops::output_metadata(&ws, &options)?;
-            cargo::print_json(&result);
-            Ok(())
-        }
-        ("new", Some(args)) => {
-            let opts = args.new_options()?;
-            ops::new(&opts, config)?;
-            let path = args.value_of("path").unwrap();
-            config.shell().status("Created", format!("{} `{}` project", opts.kind, path))?;
-            Ok(())
-        }
-        ("owner", Some(args)) => {
-            let registry = args.registry(config)?;
-            let opts = ops::OwnersOptions {
-                krate: args.value_of("crate").map(|s| s.to_string()),
-                token: args.value_of("token").map(|s| s.to_string()),
-                index: args.value_of("index").map(|s| s.to_string()),
-                to_add: args.values_of("add")
-                    .map(|xs| xs.map(|s| s.to_string()).collect()),
-                to_remove: args.values_of("remove")
-                    .map(|xs| xs.map(|s| s.to_string()).collect()),
-                list: args.is_present("list"),
-                registry,
-            };
-            ops::modify_owners(config, &opts)?;
-            Ok(())
-        }
-        ("package", Some(args)) => {
-            let ws = args.workspace(config)?;
-            ops::package(&ws, &ops::PackageOpts {
-                config,
-                verify: !args.is_present("no-verify"),
-                list: args.is_present("list"),
-                check_metadata: !args.is_present("no-metadata"),
-                allow_dirty: args.is_present("allow-dirty"),
-                target: args.target(),
-                jobs: args.jobs()?,
-                registry: None,
-            })?;
-            Ok(())
-        }
-        ("pkgid", Some(args)) => {
-            let ws = args.workspace(config)?;
-            let spec = args.value_of("spec").or(args.value_of("package"));
-            let spec = ops::pkgid(&ws, spec)?;
-            println!("{}", spec);
-            Ok(())
-        }
-        ("publish", Some(args)) => {
-            let registry = args.registry(config)?;
-            let ws = args.workspace(config)?;
-            let index = args.index(config)?;
-
-            ops::publish(&ws, &ops::PublishOpts {
-                config,
-                token: args.value_of("token").map(|s| s.to_string()),
-                index,
-                verify: !args.is_present("no-verify"),
-                allow_dirty: args.is_present("allow-dirty"),
-                target: args.target(),
-                jobs: args.jobs()?,
-                dry_run: args.is_present("dry-run"),
-                registry,
-            })?;
-            Ok(())
-        }
-        ("read-manifest", Some(args)) => {
-            let root = args.root_manifest(config)?;
-            let pkg = Package::for_path(&root, config)?;
-            cargo::print_json(&pkg);
-            Ok(())
-        }
-        ("run", Some(args)) => {
-            let ws = args.workspace(config)?;
-
-            let mut compile_opts = args.compile_options_for_single_package(
-                config, CompileMode::Build,
-            )?;
-            if !args.is_present("example") && !args.is_present("bin") {
-                compile_opts.filter = ops::CompileFilter::Default {
-                    required_features_filterable: false,
-                };
-            };
-            match ops::run(&ws, &compile_opts, &values(args, "args"))? {
-                None => Ok(()),
-                Some(err) => {
-                    // If we never actually spawned the process then that sounds pretty
-                    // bad and we always want to forward that up.
-                    let exit = match err.exit {
-                        Some(exit) => exit,
-                        None => return Err(CliError::new(err.into(), 101)),
-                    };
-
-                    // If `-q` was passed then we suppress extra error information about
-                    // a failed process, we assume the process itself printed out enough
-                    // information about why it failed so we don't do so as well
-                    let exit_code = exit.code().unwrap_or(101);
-                    let is_quiet = config.shell().verbosity() == Verbosity::Quiet;
-                    Err(if is_quiet {
-                        CliError::code(exit_code)
-                    } else {
-                        CliError::new(err.into(), exit_code)
-                    })
-                }
-            }
-        }
-        ("rustc", Some(args)) => {
-            let ws = args.workspace(config)?;
-            let mode = match args.value_of("profile") {
-                Some("dev") | None => CompileMode::Build,
-                Some("test") => CompileMode::Test,
-                Some("bench") => CompileMode::Bench,
-                Some("check") => CompileMode::Check { test: false },
-                Some(mode) => {
-                    let err = format_err!("unknown profile: `{}`, use dev,
-                                   test, or bench", mode);
-                    return Err(CliError::new(err, 101));
-                }
-            };
-            let mut compile_opts = args.compile_options_for_single_package(
-                config, mode,
-            )?;
-            compile_opts.target_rustc_args = Some(values(args, "args"));
-            ops::compile(&ws, &compile_opts)?;
-            Ok(())
-        }
-        ("rustdoc", Some(args)) => {
-            let ws = args.workspace(config)?;
-            let mut compile_opts = args.compile_options_for_single_package(
-                config, CompileMode::Doc { deps: false },
-            )?;
-            compile_opts.target_rustdoc_args = Some(values(args, "args"));
-            let doc_opts = ops::DocOptions {
-                open_result: args.is_present("open"),
-                compile_opts,
-            };
-            ops::doc(&ws, &doc_opts)?;
-            Ok(())
-        }
-        ("search", Some(args)) => {
-            let registry = args.registry(config)?;
-            let index = args.index(config)?;
-            let limit = args.value_of_u32("limit")?;
-            let limit = min(100, limit.unwrap_or(10));
-            let query: Vec<&str> = args.values_of("query").unwrap_or_default().collect();
-            let query: String = query.join("+");
-            ops::search(&query, config, index, limit, registry)?;
-            Ok(())
-        }
-        ("test", Some(args)) => {
-            let ws = args.workspace(config)?;
-
-            let mut compile_opts = args.compile_options(config, CompileMode::Test)?;
-            let doc = args.is_present("doc");
-            if doc {
-                compile_opts.mode = ops::CompileMode::Doctest;
-                compile_opts.filter = ops::CompileFilter::new(true,
-                                                              Vec::new(), false,
-                                                              Vec::new(), false,
-                                                              Vec::new(), false,
-                                                              Vec::new(), false,
-                                                              false);
-            }
-
-            let ops = ops::TestOptions {
-                no_run: args.is_present("no-run"),
-                no_fail_fast: args.is_present("no-fail-fast"),
-                only_doc: doc,
-                compile_opts,
-            };
-
-            // TESTNAME is actually an argument of the test binary, but it's
-            // important so we explicitly mention it and reconfigure
-            let mut test_args = vec![];
-            test_args.extend(args.value_of("TESTNAME").into_iter().map(|s| s.to_string()));
-            test_args.extend(args.values_of("args").unwrap_or_default().map(|s| s.to_string()));
-
-            let err = ops::run_tests(&ws, &ops, &test_args)?;
-            return match err {
-                None => Ok(()),
-                Some(err) => {
-                    Err(match err.exit.as_ref().and_then(|e| e.code()) {
-                        Some(i) => CliError::new(format_err!("{}", err.hint(&ws)), i),
-                        None => CliError::new(err.into(), 101),
-                    })
-                }
-            };
-        }
-        ("uninstall", Some(args)) => {
-            let root = args.value_of("root");
-            let specs = args.values_of("spec").unwrap_or_default().collect();
-            ops::uninstall(root, specs, &values(args, "bin"), config)?;
-            Ok(())
-        }
-        ("update", Some(args)) => {
-            let ws = args.workspace(config)?;
-
-            let update_opts = ops::UpdateOptions {
-                aggressive: args.is_present("aggressive"),
-                precise: args.value_of("precise"),
-                to_update: values(args, "package"),
-                config,
-            };
-            ops::update_lockfile(&ws, &update_opts)?;
-            Ok(())
-        }
-        ("verify-project", Some(args)) => {
-            fn fail(reason: &str, value: &str) -> ! {
-                let mut h = HashMap::new();
-                h.insert(reason.to_string(), value.to_string());
-                cargo::print_json(&h);
-                process::exit(1)
-            }
-
-            let mut contents = String::new();
-            let filename = match args.root_manifest(config) {
-                Ok(filename) => filename,
-                Err(e) => fail("invalid", &e.to_string()),
-            };
-
-            let file = File::open(&filename);
-            match file.and_then(|mut f| f.read_to_string(&mut contents)) {
-                Ok(_) => {}
-                Err(e) => fail("invalid", &format!("error reading file: {}", e))
-            };
-            if contents.parse::<toml::Value>().is_err() {
-                fail("invalid", "invalid-format");
-            }
-
-            let mut h = HashMap::new();
-            h.insert("success".to_string(), "true".to_string());
-            cargo::print_json(&h);
-            Ok(())
-        }
-        ("version", _) => {
-            println!("{}", cargo::version());
-            Ok(())
-        }
-        ("yank", Some(args)) => {
-            let registry = args.registry(config)?;
-
-            ops::yank(config,
-                      args.value_of("crate").map(|s| s.to_string()),
-                      args.value_of("vers").map(|s| s.to_string()),
-                      args.value_of("token").map(|s| s.to_string()),
-                      args.value_of("index").map(|s| s.to_string()),
-                      args.is_present("undo"),
-                      registry)?;
-            Ok(())
-        }
-        (cmd, Some(args)) => {
-            if let Some(mut alias) = super::aliased_command(config, cmd)? {
-                alias.extend(args.values_of("").unwrap_or_default().map(|s| s.to_string()));
-                let args = cli()
-                    .setting(AppSettings::NoBinaryName)
-                    .get_matches_from_safe(alias)?;
-                return execute_subcommand(config, args);
-            }
-            let mut ext_args: Vec<&str> = vec![cmd];
-            ext_args.extend(args.values_of("").unwrap_or_default());
-            super::execute_external_subcommand(config, cmd, &ext_args)
-        }
-        _ => Ok(())
+    if let Some(exec) = commands::builtin_exec(cmd) {
+        return exec(config, args);
     }
+
+    if let Some(mut alias) = super::aliased_command(config, cmd)? {
+        alias.extend(args.values_of("").unwrap_or_default().map(|s| s.to_string()));
+        let args = cli()
+            .setting(AppSettings::NoBinaryName)
+            .get_matches_from_safe(alias)?;
+        return execute_subcommand(config, args);
+    }
+    let mut ext_args: Vec<&str> = vec![cmd];
+    ext_args.extend(args.values_of("").unwrap_or_default());
+    super::execute_external_subcommand(config, cmd, &ext_args)
 }
