+++ src/bin/commands/clean.rs
@@ -1,5 +1,7 @@
 use command_prelude::*;
 
+use cargo::ops::{self, CleanOptions};
+
 pub fn cli() -> App {
     subcommand("clean")
         .about("Remove artifacts that cargo has generated in the past")
@@ -16,4 +18,16 @@ which indicates which package's artifacts should be cleaned out. If it is not
 given, then all packages' artifacts are removed. For more information on SPEC
 and its format, see the `cargo help pkgid` command.
 ")
+}
+
+pub fn exec(config: &mut Config, args: &ArgMatches) -> CliResult {
+    let ws = args.workspace(config)?;
+    let opts = CleanOptions {
+        config,
+        spec: values(args, "package"),
+        target: args.target(),
+        release: args.is_present("release"),
+    };
+    ops::clean(&ws, &opts)?;
+    Ok(())
 }
