+++ src/bin/commands/generate_lockfile.rs
@@ -1,5 +1,7 @@
 use command_prelude::*;
 
+use cargo::ops;
+
 pub fn cli() -> App {
     subcommand("generate-lockfile")
         .about("Generate the lockfile for a project")
@@ -14,4 +16,10 @@ If the lockfile is not available, then this is the equivalent of
 `cargo generate-lockfile`. A lockfile is generated and dependencies are also
 all updated.
 ")
+}
+
+pub fn exec(config: &mut Config, args: &ArgMatches) -> CliResult {
+    let ws = args.workspace(config)?;
+    ops::generate_lockfile(&ws)?;
+    Ok(())
 }
