+++ src/bin/commands/metadata.rs
@@ -1,5 +1,8 @@
 use command_prelude::*;
 
+use cargo::ops::{self, OutputMetadataOptions};
+use cargo::print_json;
+
 pub fn cli() -> App {
     subcommand("metadata")
         .about("Output the resolved dependencies of a project, \
@@ -15,4 +18,31 @@ pub fn cli() -> App {
             opt("format-version", "Format version")
                 .value_name("VERSION").possible_value("1")
         )
+}
+
+pub fn exec(config: &mut Config, args: &ArgMatches) -> CliResult {
+    let ws = args.workspace(config)?;
+
+    let version = match args.value_of("format-version") {
+        None => {
+            config.shell().warn("\
+                        please specify `--format-version` flag explicitly \
+                        to avoid compatibility problems"
+            )?;
+            1
+        }
+        Some(version) => version.parse().unwrap(),
+    };
+
+    let options = OutputMetadataOptions {
+        features: values(args, "features"),
+        all_features: args.is_present("all-features"),
+        no_default_features: args.is_present("no-default-features"),
+        no_deps: args.is_present("no-deps"),
+        version,
+    };
+
+    let result = ops::output_metadata(&ws, &options)?;
+    print_json(&result);
+    Ok(())
 }
