+++ src/bin/commands/package.rs
@@ -1,5 +1,8 @@
 use command_prelude::*;
 
+use cargo::ops::{self, PackageOpts};
+
+
 pub fn cli() -> App {
     subcommand("package")
         .about("Assemble the local package into a distributable tarball")
@@ -10,4 +13,19 @@ pub fn cli() -> App {
         .arg_target_triple("Build for the target triple")
         .arg_manifest_path()
         .arg_jobs()
+}
+
+pub fn exec(config: &mut Config, args: &ArgMatches) -> CliResult {
+    let ws = args.workspace(config)?;
+    ops::package(&ws, &PackageOpts {
+        config,
+        verify: !args.is_present("no-verify"),
+        list: args.is_present("list"),
+        check_metadata: !args.is_present("no-metadata"),
+        allow_dirty: args.is_present("allow-dirty"),
+        target: args.target(),
+        jobs: args.jobs()?,
+        registry: None,
+    })?;
+    Ok(())
 }
