+++ src/bin/commands/uninstall.rs
@@ -1,5 +1,7 @@
 use command_prelude::*;
 
+use cargo::ops;
+
 pub fn cli() -> App {
     subcommand("uninstall")
         .about("Remove a Rust binary")
@@ -18,4 +20,11 @@ specify which crate should be uninstalled. By default all binaries are
 uninstalled for a crate but the `--bin` and `--example` flags can be used to
 only uninstall particular binaries.
 ")
+}
+
+pub fn exec(config: &mut Config, args: &ArgMatches) -> CliResult {
+    let root = args.value_of("root");
+    let specs = args.values_of("spec").unwrap_or_default().collect();
+    ops::uninstall(root, specs, &values(args, "bin"), config)?;
+    Ok(())
 }
