+++ src/bin/commands/yank.rs
@@ -1,5 +1,7 @@
 use command_prelude::*;
 
+use cargo::ops;
+
 pub fn cli() -> App {
     subcommand("yank")
         .about("Remove a pushed crate from the index")
@@ -20,4 +22,17 @@ Note that existing crates locked to a yanked version will still be able to
 download the yanked version to use it. Cargo will, however, not allow any new
 crates to be locked to any yanked version.
 ")
+}
+
+pub fn exec(config: &mut Config, args: &ArgMatches) -> CliResult {
+    let registry = args.registry(config)?;
+
+    ops::yank(config,
+              args.value_of("crate").map(|s| s.to_string()),
+              args.value_of("vers").map(|s| s.to_string()),
+              args.value_of("token").map(|s| s.to_string()),
+              args.value_of("index").map(|s| s.to_string()),
+              args.is_present("undo"),
+              registry)?;
+    Ok(())
 }
